+++ CHANGELOG.md
@@ -1939,3 +1939,4 @@
 - [2026-06-12 14:41] test(storage): implement widget integration tests for storage
 - [2026-06-12 14:46] chore(security): update dependencies and cleanup unused imports in security
 - [2026-06-12 16:15] chore(analytics): update dependencies and cleanup unused imports in analytics
+- [2026-06-12 16:18] chore(navigation): update dependencies and cleanup unused imports in navigation